changes lab04b v3

+++ lab04b/agg/src/main/scala/agg.scala
@@ -39,7 +39,7 @@ object agg {
       .withWatermark("timestamp", "1 hour")
       .groupBy(window($"timestamp", "1 hour","1 hour"))
       .agg(sum(when('event_type === "buy",col("item_price"))).as("revenue"),count('event_type === "buy").as("purchases"),count(col("uid").isNotNull).as("visitors"))
-      .select(col("window.start").cast("long").as("start_ts"),col("window.end").cast("long").as("end_ts"),'revenue,'visitors,'purchases, ('revenue.cast("double")/'purchases.cast("double")).as("aov"))
+      .select(col("window.start").cast("long").as("start_ts"),col("window.end").cast("long").as("end_ts"),'revenue,'visitors,'purchases, ('revenue/'purchases).as("aov"))
       .toJSON
       .writeStream
       .trigger(Trigger.ProcessingTime("10 seconds"))
